Refactor util and create core class - Remove utils module and its contents (file_helper.py) - Create the core class TrtLogFile, available in trt_log_file.py - Move the open_file functionality from file_helper to TrtLogFile
